+++ gantry/entrypoint.sh
@@ -149,14 +149,17 @@ function ensure_conda {
     fi
 
     if [[ -z "$GANTRY_CONDA_INITIALIZED" ]]; then
-        log_info "Configuring conda for shell environment..."
+        log_info "Configuring $(conda --version) for shell environment..."
         # Initialize conda for bash.
         # See https://stackoverflow.com/a/58081608/4151392
         eval "$(command conda 'shell.bash' 'hook' 2> /dev/null)"
 
         # Accept TOS for default channels.
-        capture_logs "conda_tos_accept_main.log" conda tos accept --override-channels --channel https://repo.anaconda.com/pkgs/main || return 1
-        capture_logs "conda_tos_accept_r.log" conda tos accept --override-channels --channel https://repo.anaconda.com/pkgs/r || return 1
+        # NOTE: this will fail if the conda version is too old.
+        if command conda tos &> /dev/null 2>&1; then
+            capture_logs "conda_tos_accept_main.log" conda tos accept --override-channels --channel https://repo.anaconda.com/pkgs/main
+            capture_logs "conda_tos_accept_r.log" conda tos accept --override-channels --channel https://repo.anaconda.com/pkgs/r
+        fi
 
         GANTRY_CONDA_INITIALIZED="1"
         log_info "Done."